requirements.txt
streamlit
playwright
Pillow
psutil
playwright-stealth
streamlit-autorefresh



